refactor: rename api package and CLI from freehold to marrow (#67)
Rename the backend Python package api/freehold/ to api/marrow/ and
the CLI entrypoint freehold to marrow. Update all imports across
main.py, tests, and alembic/env.py, plus pyproject.toml package
name, scripts entry, and setuptools include pattern.

Scope limited to package/CLI rename per issue #67. Related renames
(DB identifiers, session cookie, export bundle prefix) are tracked
separately in #68, #69, #70.
