+++ api/alembic/env.py
@@ -20,7 +20,7 @@ config.set_main_option("sqlalchemy.url", os.environ["DATABASE_URL"])
 if config.config_file_name is not None:
     fileConfig(config.config_file_name)
 
-from freehold.models import Base  # noqa: E402
+from marrow.models import Base  # noqa: E402
 
 target_metadata = Base.metadata
