+++ api/main.py
@@ -4,4 +4,4 @@ Run from the api/ directory:
     uvicorn main:app --reload
 """
 
-from freehold.app import app  # noqa: F401  re-exported for uvicorn
+from marrow.app import app  # noqa: F401  re-exported for uvicorn
