+++ api/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "freehold-api"
+name = "marrow-api"
 version = "0.1.0"
 requires-python = ">=3.11"
 dependencies = [
@@ -18,7 +18,7 @@ dependencies = [
 ]
 
 [project.scripts]
-freehold = "freehold.cli:app"
+marrow = "marrow.cli:app"
 
 [project.optional-dependencies]
 dev = [
@@ -29,7 +29,7 @@ dev = [
 ]
 
 [tool.setuptools.packages.find]
-include = ["freehold*"]
+include = ["marrow*"]
 
 [tool.ruff]
 target-version = "py311"
